some initial configuration

+++ tools/makesymlinks.sh
@@ -8,7 +8,7 @@
 
 dir=~/dotfiles                    # dotfiles directory
 olddir=~/dotfiles_old             # old dotfiles backup directory
-files=".bashrc .tmux.conf .gitconfig  .vimrc"    # list of files/folders to symlink in homedir
+files=".bashrc .tmuxp/ .tmux.conf .gitconfig  .vimrc"    # list of files/folders to symlink in homedir
 
 ##########
 
@@ -27,5 +27,12 @@ for file in $files; do
     echo "Moving any existing dotfiles from ~ to $olddir"
     mv ~/$file ~/dotfiles_old/
     echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/$file
+    if [ -e ~/$file ]
+    then
+	echo "$file has a symlink already"
+    else
+    	ln -s $dir/$file ~/$file
+    fi
 done
+
+
